Conflicts:
	app/Http/Controllers/LocatarioController.php
	app/Http/Controllers/LocatoreController.php
	app/Models/Catalog.php
	app/User.php
	database/migrations/2022_05_19_174521_contracts_migration.php
	database/migrations/2022_06_01_130033_options_migration.php
	database/seeds/DatabaseSeeder.php
	public/assets/css/style.css
	public/assets/js/filtersCatalog.js
	routes/web.php

+++ app/Http/Controllers/LocatoreController.php
@@ -26,11 +26,22 @@ class LocatoreController extends Controller {
                         ->with('idLarioList', $this->offerListModel->getUserByOffer(Auth::id()))
                         ->with('optionList', $this->offerListModel->getAllOption());
     }
+    
+    public function acceptOffer($opt_id) {
+        
+        $this->offerListModel->setAcceptById($opt_id);
+        return redirect()->action('LocatoreController@showOfferList');
+    }
 
     public function deleteOffer($offerId) {
-
-        Offer::where('offerta_id', $offerId)->delete();
-        return redirect()->action('LocatoreController@showOfferList');
+        
+        $autenticati = Offer::where('offerta_id', $offerId)->value('user_id');
+        
+        if ( Auth::id() == $autenticati){
+         Offer::where('offerta_id', $offerId)->delete();
+         return redirect()->action('LocatoreController@showOfferList');
+        }
+        return redirect()->action('UserController@index');
     }
 
     public function insertOffer() {
